# Conflicts:
#	tests/integration/cases/float_special_values/Perl@v5_36.pl
#	tests/integration/cases/float_special_values/Perl_combined@v5_36.pl
#	tests/integration/cases/float_special_values/Perl_float_format_fixed_v@v5_36.pl
#	tests/integration/cases/float_special_values/Perl_float_format_scientific_v@v5_36.pl
#	tests/integration/cases/float_special_values/Perl_numeric_separator_underscore_float_v@v5_36.pl

+++ tests/integration/cases/list_mixed_int_widths/Go_combined@v1_18.go
@@ -1,11 +1,11 @@
 package main
 
 func main() {
-my_data := []int{
+my_data := []int64{
 	1,
 	1099511627776,
 }
-my_data = []int{
+my_data = []int64{
 	1,
 	1099511627776,
 }
