+++ tests/integration/cases/map_mixed_int_widths/Go_declaration_style_var@v1_18.go
@@ -1,7 +1,7 @@
 package main
 
 func main() {
-var my_data = map[string]int{
+var my_data = map[string]int64{
 	"a": 1,
 	"b": 1099511627776,
 }
