+++ tests/integration/cases/map_mixed_int_widths/Go_type_hints_safe@v1_18.go
@@ -1,7 +1,7 @@
 package main
 
 func main() {
-my_data := map[string]int{
+my_data := map[string]int64{
 	"a": 1,
 	"b": 1099511627776,
 }
